更新代码与构建脚本
1. 修复 Windows 下构造函数未启用的问题；
2. 完善跨平台变量类型；
3. 更新测试构建脚本，使更合理

+++ include/sirius/custom/custom_thread.h
@@ -11,23 +11,23 @@ extern "C" {
     defined(__OpenBSD__) || defined(__NetBSD__)
 #include <sys/syscall.h>
 #include <unistd.h>
-static inline unsigned long long _custom_thread_id() {
-  return (unsigned long long)syscall(SYS_gettid);
+static inline uint64_t _custom_thread_id() {
+  return (uint64_t)syscall(SYS_gettid);
 }
 #elif defined(_WIN32)
-static inline unsigned long long _custom_thread_id() {
-  return (unsigned long long)GetCurrentThreadId();
+static inline uint64_t _custom_thread_id() {
+  return (uint64_t)GetCurrentThreadId();
 }
 #elif defined(__APPLE__) && defined(__MACH__)
-static inline unsigned long long _custom_thread_id() {
-  unsigned long long thread_id;
+static inline uint64_t _custom_thread_id() {
+  uint64_t thread_id;
   if (unlikely(pthread_threadid_np(NULL, &thread_id) != 0))
     return 0;
   return thread_id;
 }
 #elif defined(sun) || defined(__sun)
-static inline unsigned long long _custom_thread_id() {
-  return (unsigned long long)thr_self();
+static inline uint64_t _custom_thread_id() {
+  return (uint64_t)thr_self();
 }
 #else
 #define _custom_thread_id() 0
